% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airm.R
\name{airm_log}
\alias{airm_log}
\title{Compute the AIRM Logarithm}
\usage{
airm_log(sigma, lambda)
}
\arguments{
\item{sigma}{A symmetric positive-definite matrix of class \code{dppMatrix}, representing the reference point. # nolint: line_length_linter}

\item{lambda}{A symmetric positive-definite matrix of class \code{dppMatrix}, representing the target point. # nolint: line_length_linter}
}
\value{
A symmetric matrix of class \code{dspMatrix}, representing the tangent space image of \code{lambda} at \code{sigma}. # nolint: line_length_linter
}
\description{
This function computes the Riemannian logarithmic map for the Affine-Invariant Riemannian Metric (AIRM). # nolint: line_length_linter
}
\examples{
\dontrun{
sigma <- diag(2) |> Matrix::nearPD()$mat |> Matrix::pack()
lambda <- diag(c(2, 3)) |> Matrix::nearPD()$mat |> Matrix::pack()
airm_log(sigma, lambda)
}
}
