% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bures-wasserstein.R
\name{bures_wasserstein_unvec}
\alias{bures_wasserstein_unvec}
\title{Compute the Bures-Wasserstein Inverse Vectorization}
\usage{
bures_wasserstein_unvec(sigma, w)
}
\arguments{
\item{sigma}{A symmetric positive-definite matrix of class \code{dppMatrix}}

\item{w}{A numeric vector representing the vectorized tangent image}
}
\value{
A symmetric matrix of class \code{dspMatrix}
}
\description{
Compute the Bures-Wasserstein Inverse Vectorization
}
