% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bures-wasserstein.R
\name{bures_wasserstein_vec}
\alias{bures_wasserstein_vec}
\title{Compute the Bures-Wasserstein Vectorization}
\usage{
bures_wasserstein_vec(sigma, v)
}
\arguments{
\item{sigma}{A symmetric positive-definite matrix of class \code{dppMatrix}}

\item{v}{A symmetric matrix of class \code{dspMatrix}}
}
\value{
A numeric vector representing the vectorized tangent image
}
\description{
Compute the Bures-Wasserstein Vectorization
}
