% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{default_ref_pt}
\alias{default_ref_pt}
\title{Default reference point}
\usage{
default_ref_pt(p)
}
\arguments{
\item{p}{the dimension}
}
\value{
A diagonal matrix of the desired dimension
}
\description{
Default reference point
}
