% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{dexp}
\alias{dexp}
\title{Differential of Matrix Exponential Map}
\usage{
dexp(a, x)
}
\arguments{
\item{a}{A symmetric matrix of class dspMatrix}

\item{x}{A symmetric matrix representing tangent vector of class dspMatrix}
}
\value{
A positive definite symmetric matrix representing the differential located at a and evaluated at x, of class dppMatrix # nolint: line_length_linter
}
\description{
Computes the differential of the matrix exponential map located at a point a, evaluated at x # nolint: line_length_linter
}
