% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{dlog}
\alias{dlog}
\title{Differential of Matrix Logarithm Map}
\usage{
dlog(sigma, h)
}
\arguments{
\item{sigma}{A symmetric positive definite matrix of class dspMatrix}

\item{H}{A symmetric matrix representing tangent vector of class dsyMatrix}
}
\value{
A symmetric matrix representing the differential evaluated at H of class dsyMatrix # nolint: line_length_linter
}
\description{
Computes the differential of the matrix logarithm map at a point Sigma, evaluated at H # nolint: line_length_linter
}
