% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{half_underscore}
\alias{half_underscore}
\title{Half-underscore operation for use in the log-Cholesky metric}
\usage{
half_underscore(x)
}
\arguments{
\item{x}{A symmetric matrix (object of class dsyMatrix)}
}
\value{
The strictly lower triangular part of the matrix, plus half its diagonal part #nolint: line_length_linter
}
\description{
Half-underscore operation for use in the log-Cholesky metric
}
