% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{id_matr}
\alias{id_matr}
\title{Create an Identity Matrix}
\usage{
id_matr(sigma)
}
\arguments{
\item{sigma}{A matrix.}
}
\value{
An identity matrix of the same dimensions as \code{sigma}.
}
\description{
Create an Identity Matrix
}
