% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_cholesky.R
\name{log_cholesky_unvec}
\alias{log_cholesky_unvec}
\title{Compute the Log-Cholesky Inverse Vectorization}
\usage{
log_cholesky_unvec(sigma, w)
}
\arguments{
\item{sigma}{A symmetric positive-definite matrix of class \code{dppMatrix}}

\item{w}{A numeric vector representing the vectorized tangent image}
}
\value{
A symmetric matrix of class \code{dspMatrix}
}
\description{
Compute the Log-Cholesky Inverse Vectorization
}
