% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_cholesky.R
\name{log_cholesky_vec}
\alias{log_cholesky_vec}
\title{Compute the Log-Cholesky Vectorization}
\usage{
log_cholesky_vec(sigma, v)
}
\arguments{
\item{sigma}{A symmetric positive-definite matrix of class \code{dppMatrix}}

\item{v}{A symmetric matrix of class \code{dspMatrix}}
}
\value{
A numeric vector representing the vectorized tangent image
}
\description{
Compute the Log-Cholesky Vectorization
}
