% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_euclidean.R
\name{log_euclidean_exp}
\alias{log_euclidean_exp}
\title{Compute the Log-Euclidean Exponential}
\usage{
log_euclidean_exp(ref_pt, v)
}
\arguments{
\item{ref_pt}{A reference point.}

\item{v}{A tangent vector to be mapped back to the manifold at \code{ref_pt}. # nolint: line_length_linter}
}
\value{
The point on the manifold corresponding to the tangent vector at \code{ref_pt}. # nolint: line_length_linter
}
\description{
This function computes the Euclidean exponential map.
}
