% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_euclidean.R
\name{log_euclidean_log}
\alias{log_euclidean_log}
\title{Compute the Log-Euclidean Logarithm}
\usage{
log_euclidean_log(sigma, lambda)
}
\arguments{
\item{sigma}{A reference point.}

\item{lambda}{A point on the manifold.}
}
\value{
The tangent space image of \code{lambda} at \code{sigma}.
}
\description{
Compute the Log-Euclidean Logarithm
}
