% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_euclidean.R
\name{log_euclidean_unvec}
\alias{log_euclidean_unvec}
\title{Compute the Inverse Vectorization (Euclidean)}
\usage{
log_euclidean_unvec(sigma, w)
}
\arguments{
\item{sigma}{A symmetric matrix.}

\item{w}{A numeric vector, representing the vectorized tangent image.}
}
\value{
A symmetric matrix, representing the tangent vector.
}
\description{
Converts a vector back into a tangent matrix relative to a reference point using Euclidean metric. # nolint: line_length_linter
}
