% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_euclidean.R
\name{log_euclidean_vec}
\alias{log_euclidean_vec}
\title{Vectorize at Identity Matrix (Euclidean)}
\usage{
log_euclidean_vec(sigma, v)
}
\arguments{
\item{sigma}{A symmetric matrix.}

\item{v}{A vector.}
}
\value{
A numeric vector, representing the vectorized tangent image.
}
\description{
Converts a symmetric matrix into a vector representation. # nolint: line_length_linter
}
