% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics}
\alias{metrics}
\alias{airm}
\alias{log_euclidean}
\alias{euclidean}
\alias{log_cholesky}
\alias{bures_wasserstein}
\title{Pre-configured Riemannian metrics for SPD matrices}
\format{
Objects of class \code{rmetric} containing four functions:
\describe{
\item{log}{Computes the Riemannian logarithm}
\item{exp}{Computes the Riemannian exponential}
\item{vec}{Performs vectorization}
\item{unvec}{Performs inverse vectorization}
}

An object of class \code{rmetric} of length 4.

An object of class \code{rmetric} of length 4.

An object of class \code{rmetric} of length 4.

An object of class \code{rmetric} of length 4.

An object of class \code{rmetric} of length 4.
}
\usage{
airm

log_euclidean

euclidean

log_cholesky

bures_wasserstein
}
\description{
Ready-to-use metric objects for various Riemannian geometries on the manifold
of symmetric positive definite matrices.
}
\examples{
data(airm)

# Create two SPD matrices
sigma <- Matrix(c(2.0, 0.5, 0.5, 3.0), nrow = 2) |>
    nearPD() |> _$mat |> pack()
lambda <- Matrix(c(1.5, 0.3, 0.3, 2.5), nrow = 2) |>
    nearPD() |> _$mat |> pack()

# Use the metric
result <- airm$log(sigma, lambda)
}
\keyword{datasets}
