% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{safe_logm}
\alias{safe_logm}
\title{Wrapper for the matrix logarithm}
\usage{
safe_logm(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Its matrix logarithm
}
\description{
Wrapper for the matrix logarithm
}
