% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bures-wasserstein.R
\name{solve_lyapunov}
\alias{solve_lyapunov}
\title{Solve the Lyapunov Equation}
\usage{
solve_lyapunov(p, v)
}
\arguments{
\item{p}{A symmetric positive-definite matrix of class \code{dppMatrix}.}

\item{v}{A symmetric matrix of class \code{dspMatrix}.}
}
\value{
A symmetric matrix of class \code{dspMatrix}, representing the solution L.
}
\description{
Solves the Lyapunov equation L P + P L = V for L.
}
\keyword{internal}
