% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_cholesky.R
\name{spd_isometry_from_identity}
\alias{spd_isometry_from_identity}
\title{Reverse isometry from tangent space at identity to tangent space at P}
\usage{
spd_isometry_from_identity(sigma, v)
}
\arguments{
\item{sigma}{A symmetric positive-definite matrix of class \code{dppMatrix}}

\item{v}{A symmetric matrix of class \code{dspMatrix}}
}
\value{
A symmetric matrix of class \code{dspMatrix}
}
\description{
Reverse isometry from tangent space at identity to tangent space at P
}
