% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_cholesky.R
\name{spd_isometry_to_identity}
\alias{spd_isometry_to_identity}
\title{Isometry from tangent space at P to tangent space at identity}
\usage{
spd_isometry_to_identity(sigma, v)
}
\arguments{
\item{sigma}{A symmetric positive-definite matrix of class \code{dppMatrix}}

\item{v}{A symmetric matrix of class \code{dspMatrix}}
}
\value{
A symmetric matrix of class \code{dspMatrix}
}
\description{
Isometry from tangent space at P to tangent space at identity
}
