% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{validate_conns}
\alias{validate_conns}
\title{Validate Connections}
\usage{
validate_conns(conns, tan_imgs, vec_imgs, centered)
}
\arguments{
\item{conns}{List of connection matrices.}

\item{tan_imgs}{List of tangent images.}

\item{vec_imgs}{Matrix of vector images.}

\item{centered}{Logical indicating if the data is centered.}
}
\value{
None. Throws an error if the validation fails.
}
\description{
Validates the connections input.
}
