% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{validate_log_args}
\alias{validate_log_args}
\title{Validate arguments for Riemannian logarithms}
\usage{
validate_log_args(sigma, lambda)
}
\arguments{
\item{sigma}{A dppMatrix object}

\item{lambda}{A dppMatrix object}
}
\value{
None
}
\description{
Validate arguments for Riemannian logarithms
}
\details{
Error if sigma and lambda are not of the same dimensions
}
