% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{validate_metric}
\alias{validate_metric}
\title{Validate Metric}
\usage{
validate_metric(metric)
}
\arguments{
\item{metric}{The metric to validate.}
}
\value{
None. Throws an error if the metric is NULL.
}
\description{
Validates that the metric is not NULL.
}
