% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{validate_tan_imgs}
\alias{validate_tan_imgs}
\title{Validate Tangent Images}
\usage{
validate_tan_imgs(tan_imgs, vec_imgs, centered)
}
\arguments{
\item{tan_imgs}{List of tangent images.}

\item{vec_imgs}{List of vector images.}

\item{centered}{Logical indicating if the data is centered.}
}
\value{
None. Throws an error if the validation fails.
}
\description{
Validates the tangent images input.
}
