% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{validate_unvec_args}
\alias{validate_unvec_args}
\title{Validate arguments for inverse vectorization}
\usage{
validate_unvec_args(sigma, w)
}
\arguments{
\item{sigma}{A dppMatrix object}

\item{w}{A numeric vector}
}
\value{
None
}
\description{
Validate arguments for inverse vectorization
}
\details{
Error if the dimensionalities don't match
}
