% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{validate_vec_args}
\alias{validate_vec_args}
\title{Validate arguments for vectorization}
\usage{
validate_vec_args(sigma, v)
}
\arguments{
\item{sigma}{A dppMatrix object}

\item{v}{A dspMatrix object}
}
\value{
None
}
\description{
Validate arguments for vectorization
}
\details{
Error if sigma and v are not of the same dimensions
}
