% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{validate_vec_imgs}
\alias{validate_vec_imgs}
\title{Validate Vector Images}
\usage{
validate_vec_imgs(vec_imgs, centered)
}
\arguments{
\item{vec_imgs}{List of vector images.}

\item{centered}{Logical indicating if the data is centered.}
}
\value{
None. Throws an error if the validation fails.
}
\description{
Validates the vector images input.
}
